+++ app/models/competition.rb
@@ -1,6 +1,6 @@
 class Competition < ApplicationRecord
   ## relations
-  has_many :category_results, dependent: :destroy
+  has_many :results, dependent: :destroy
   has_many :scores, dependent: :destroy
 
   ## validations
@@ -17,7 +17,7 @@ class Competition < ApplicationRecord
   end
   ## updater
   def clean
-    category_results.map(&:destroy)
+    results.map(&:destroy)
     scores.map(&:destroy)
   end
   def update!
@@ -40,8 +40,8 @@ class Competition < ApplicationRecord
       ## categories
       parsed[:categories].each do |category, cat_item|
         next unless Category.accept?(category)
-        Parsers.parser(:category_result, parser_type.to_sym).parse(cat_item[:result_url]).each do |cr_parsed|
-          category_results.create!(category: category) do |cr|
+        Parsers.parser(:result, parser_type.to_sym).parse(cat_item[:result_url]).each do |cr_parsed|
+          results.create!(category: category) do |cr|
             cr.update!(cr_parsed)
             puts cr.summary
           end
@@ -51,14 +51,14 @@ class Competition < ApplicationRecord
         parsed[:segments][category].each do |segment, seg_item|
           Parser::ScoreParser.new.parse(seg_item[:score_url]).each do |sc_parsed|
             scores.create!(category: category, segment: segment) do |score|
-              cr_rels = category_results.where(category: category)
+              cr_rels = results.where(category: category)
               relevant_cr =
                 cr_rels.find_by_skater_name(sc_parsed[:skater_name]) ||
                 cr_rels.where(category: category).find_by_segment_ranking(segment, sc_parsed[:ranking]) ||
                 raise("no relevant category results for %<skater_name>s %<segment>s#%<ranking>d" % sc_parsed.merge(segment: segment))
                       
               score.attributes = {
-                category_result: relevant_cr,
+                result: relevant_cr,
                 skater: relevant_cr.skater,
                 date: seg_item[:date],
               }